python-ms/historical_processing/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pydantic-settings==2.1.0
redis==5.0.1

# HTTP Client
httpx==0.26.0

# Database/Storage
psycopg2-binary==2.9.9

# Utilities
python-multipart==0.0.6
python-dateutil==2.9.0.post0

# Development extras (optional)
pytest==8.0.0  
numpy>=1.24.0
